requirements.txt
Flask
openai
pymongo
webvtt-py